fix(ci): resolve all clippy -D warnings and test feature gate failures
Clippy fixes across src, benches, and tests (zero errors after this):
- Replace manual div_ceil/is_multiple_of with stdlib methods
- Derive Default for enums/structs where applicable (#[default] variant)
- Remove Clone on Copy types (Option<ModelTier>, etc.)
- Fix too_many_arguments: #[allow] where params are legitimately distinct
- Fix needless closures, unneeded returns, collapsible ifs
- Fix map_or -> is_some_and, as_ref.map noop, lifetime elision issues
- Fix indexing loops to use enumerate(), array literals over vec!
- Gate dead code (simple_tokenize) behind correct feature

Test feature gate fixes:
- Add #![cfg(feature = "advanced")] to simd_matmul_test.rs,
  speculative_test.rs, simd_neon_test.rs
- Gate FlashAttnGpuConfig/MultiGpu* imports in gpu_v2_test.rs

Root .cargo/config.toml:
- Verify [build] target = x86_64-pc-windows-msvc removed (cross-compile
  fix for Ubuntu/macOS CI — this was the second config.toml that was
  still forcing Windows MSVC target on non-Windows runners)

+++ core-runtime/tests/tier2_onnx_classification_test.rs
@@ -296,7 +296,7 @@ async fn test_concurrent_classification_requests() {
     // Validate all requests completed
     assert_eq!(
         results.len(),
-        num_concurrent as usize,
+        num_concurrent,
         "All requests should complete"
     );
